feat(tooling-explorer): Move CustomRPCInput component to own file
