+++ apps/explorer/src/components/network/WrappedNetworkSelect.tsx
@@ -5,16 +5,14 @@
 import cx from 'clsx';
 import { useIotaClientQuery } from '@iota/dapp-kit';
 import { useContext, useEffect, useRef, useState } from 'react';
-import { z } from 'zod';
 
 import { NetworkContext } from '~/contexts';
-import { type NetworkOption } from '~/components/ui';
+import { CustomRPCInput, type NetworkOption } from '~/components/ui';
 import { ampli } from '~/lib/utils';
 import { type NetworkId, getAllNetworks } from '@iota/iota-sdk/client';
 import { Button, ButtonSize, ButtonType, Dropdown, ListItem } from '@iota/apps-ui-kit';
 import { ArrowLeft, ArrowDown, CheckmarkFilled } from '@iota/ui-icons';
 import { Transition } from '@headlessui/react';
-import { useZodForm } from '@iota/core';
 
 export function NetworkSelector(): JSX.Element {
     const elementRef = useRef<HTMLDivElement>(null);
@@ -149,56 +147,5 @@ export function NetworkSelector(): JSX.Element {
                 </div>
             </Transition>
         </div>
-    );
-}
-
-const CustomRPCSchema = z.object({
-    url: z.string().url(),
-});
-
-interface CustomRPCInputProps {
-    value: string;
-    onChange(networkUrl: string): void;
-}
-
-function CustomRPCInput({ value, onChange }: CustomRPCInputProps): JSX.Element {
-    const { register, handleSubmit, formState } = useZodForm({
-        schema: CustomRPCSchema,
-        mode: 'all',
-        defaultValues: {
-            url: value,
-        },
-    });
-
-    const { errors, isDirty, isValid } = formState;
-
-    return (
-        <form
-            onSubmit={handleSubmit((values) => {
-                onChange(values.url);
-            })}
-            className="relative flex items-center rounded-md"
-        >
-            <input
-                {...register('url')}
-                type="text"
-                className={cx(
-                    'w-full rounded-md border p-xs outline-none',
-                    errors.url
-                        ? 'border-issue-dark text-issue-dark'
-                        : 'border-gray-65 text-gray-90',
-                )}
-            />
-
-            <div className="absolute inset-y-0 right-0 flex flex-col justify-center px-xs">
-                <button
-                    disabled={!isDirty || !isValid}
-                    type="submit"
-                    className="flex items-center justify-center rounded-full bg-gray-90 px-2 py-1 text-captionSmall font-semibold uppercase text-white transition disabled:bg-gray-45 disabled:text-gray-65"
-                >
-                    Save
-                </button>
-            </div>
-        </form>
     );
 }
