+++ apps/explorer/src/components/ui/header/index.ts
@@ -3,3 +3,4 @@
 
 export * from './NavItem';
 export * from './NetworkSelect';
+export * from './CustomRPCInput';
